requirements.txt
Django==1.9
django-crispy-forms==1.7.2
